# console/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brand Console API",
    "version": "0.1.0",
    "description": "The local Express API behind the brand console. Runs unauthenticated on http://localhost:5174. Read-only reference; the console never holds a secret because there is no secret to hold."
  },
  "servers": [{ "url": "http://localhost:5174", "description": "Local dev server" }],
...
  "paths": {
...
    "/api/discovery/inbox": {
      "get": {
        "tags": ["Discovery"],
        "summary": "The triage inbox with joined tags",
        "parameters": [
          { "name": "state", "in": "query", "required": false, "schema": { "type": "string", "enum": ["inbox", "promoted", "archived"] }, "description": "Filter by triage state." },
          { "name": "q", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Free-text search over title/summary." },
          { "name": "tags", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Comma-separated tag ids." },
          { "name": "group", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Group key." }
        ],
        "responses": {
          "200": {
            "description": "Inbox items with their tags.",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/InboxItem" } }
              }
            }
          }
        }
      }
    },
    "/api/discovery/{id}/archive": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Move an item to archived",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "The updated inbox item.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboxItem" } } }
          }
        }
      }
    },
    "/api/discovery/{id}/unarchive": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Return an item to the inbox",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "The updated inbox item.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InboxItem" } } }
          }
        }
      }
    },
    "/api/discovery/{id}/promote": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Promote a feed item to a seeded idea",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "comment": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created idea reference.",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "id": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/api/discovery/{id}/tags": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Attach a tag to a feed item",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "tagId": { "type": "string" } },
                "required": ["tagId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The attachment.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "feedItemId": { "type": "string" }, "tagId": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/{id}/tags/{tagId}": {
      "delete": {
        "tags": ["Discovery"],
        "summary": "Detach a tag from a feed item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tagId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "Detached." } }
      }
    },
...
    "/api/database/tables/{name}": {
      "get": {
        "tags": ["Database"],
        "summary": "Columns plus a filtered page of rows",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Column to sort by." },
          { "name": "dir", "in": "query", "required": false, "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": {
            "description": "The page of rows. Column filters are passed as f.<column> query params.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DbTableData" } } }
          }
        }
      }
    },
    "/api/database/query": {
      "post": {
        "tags": ["Database"],
        "summary": "Run one read-only SELECT/WITH statement",
        "description": "Executes a single read-only statement. On a rejected or failing query the handler returns HTTP 200 with an `error` string in the body rather than a non-2xx status, so callers must inspect `error`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "sql": { "type": "string" } },
                "required": ["sql"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result columns and rows, or a 200-with-error body when the statement is rejected.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DbQueryResult" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "color": { "type": "string" }
        }
      },
      "InboxItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "sourceId": { "type": "string" },
          "externalId": { "type": "string" },
          "title": { "type": "string" },
          "url": { "type": ["string", "null"] },
          "summary": { "type": ["string", "null"] },
          "publishedAt": { "type": ["integer", "null"] },
          "fetchedAt": { "type": "integer" },
          "triageState": { "type": "string", "enum": ["inbox", "promoted", "archived"] },
          "archivedAt": { "type": ["integer", "null"] },
          "promotedIdeaId": { "type": ["string", "null"] },
          "score": { "type": ["number", "null"] },
          "tags": { "type": "array", "items": { "$ref": "#/components/schemas/Tag" } }
        }
      },
...
      "DbColumn": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string" },
          "notnull": { "type": "boolean" },
          "pk": { "type": "boolean" }
        }
      },
      "DbTableData": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "label": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string" },
          "columns": { "type": "array", "items": { "$ref": "#/components/schemas/DbColumn" } },
          "rowCount": { "type": "integer" },
          "filteredCount": { "type": "integer" },
          "rows": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "limit": { "type": "integer" },
          "offset": { "type": "integer" }
        }
      },
      "DbQueryResult": {
        "type": "object",
        "properties": {
          "columns": { "type": "array", "items": { "type": "string" } },
          "rows": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "truncated": { "type": "boolean" },
          "error": { "type": "string", "description": "Present when the statement was rejected or failed (returned with HTTP 200)." }
        }
      }
    }
  }
}